Move Bool extension into file
